Renamed test_app_legacy to sample_app_legacy.
