+++ notification-hubs-sample-app-legacy/appcenter-post-clone.sh
