+++ notification-hubs-sample-app-legacy/proguard-rules.pro
