+++ notification-hubs-sample-app-legacy/src/androidTest/java/com/microsoft/notification_hubs_sample_app_legacy/ExampleInstrumentedTest.java
@@ -1,4 +1,4 @@
-package com.microsoft.notification_hubs_test_app_legacy;
+package com.microsoft.notification_hubs_sample_app_legacy;
 
 import android.content.Context;
 import androidx.test.platform.app.InstrumentationRegistry;
