+++ notification-hubs-sample-app-legacy/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.microsoft.notification_hubs_test_app_legacy">
+    package="com.microsoft.notification_hubs_sample_app_legacy">
 
     <application
         android:allowBackup="true"
@@ -8,7 +8,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".MainActivity">
+        <activity android:name="com.microsoft.notification_hubs_sample_app_legacy.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
@@ -26,12 +26,12 @@
         </service>
 
         <service
-            android:name=".RegistrationIntentService"
+            android:name="com.microsoft.notification_hubs_sample_app_legacy.RegistrationIntentService"
             android:exported="false">
         </service>
 
         <service
-            android:name=".UnregistrationIntentService"
+            android:name="com.microsoft.notification_hubs_sample_app_legacy.UnregistrationIntentService"
             android:exported="false">
         </service>
     </application>
