+++ notification-hubs-sample-app-legacy/src/main/java/com/microsoft/notification_hubs_sample_app_legacy/MainActivity.java
@@ -1,4 +1,4 @@
-package com.microsoft.notification_hubs_test_app_legacy;
+package com.microsoft.notification_hubs_sample_app_legacy;
 
 import android.content.SharedPreferences;
 import android.preference.PreferenceManager;
