+++ notification-hubs-sample-app-legacy/src/main/res/drawable/ic_launcher_background.xml
