+++ notification-hubs-sample-app-legacy/src/main/res/layout/activity_main.xml
