+++ notification-hubs-sample-app-legacy/src/main/res/mipmap-anydpi-v26/ic_launcher_round.xml
