+++ notification-hubs-sample-app-legacy/src/main/res/mipmap-hdpi/ic_launcher_round.png
