+++ notification-hubs-sample-app-legacy/src/main/res/mipmap-mdpi/ic_launcher_round.png
