+++ notification-hubs-sample-app-legacy/src/main/res/mipmap-xxhdpi/ic_launcher.png
