+++ notification-hubs-sample-app-legacy/src/main/res/mipmap-xxxhdpi/ic_launcher_round.png
