+++ notification-hubs-sample-app-legacy/src/main/res/values/colors.xml
