+++ notification-hubs-sample-app-legacy/src/main/res/values/styles.xml
