+++ notification-hubs-sample-app-legacy/src/test/java/com/microsoft/notification_hubs_sample_app_legacy/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package com.microsoft.notification_hubs_test_app_legacy;
+package com.microsoft.notification_hubs_sample_app_legacy;
 
 import org.junit.Test;
